Don't use /api/articles but only /api/protected/articles
